% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflactanator.R
\name{deflactanator}
\alias{deflactanator}
\title{Llevar a precios constantes las bases}
\usage{
deflactanator(
  data,
  mes_base,
  variables_deflactar,
  variable_mes,
  pisar_datos = F
)
}
\arguments{
\item{data}{Dataframe que se quiere deflactar/inflar}

\item{mes_base}{Mes que se quiere utilizar como base para inflar/deflactar. Formato: "YYYY-MM-DD"}

\item{variables_deflactar}{Vector con los nombres de las variables que se quieren deflactar}

\item{variable_mes}{Nombre de la variable que indica el mes en la base a deflactar. Dicha variable debe estar en formato "YYYY-MM-DD"}

\item{pisar_datos}{Valor T/F que permite pisar las variables que se deflactaron}
}
\value{
A matrix of the infile
}
\description{
Se puede inflar o deflactar los datos con la funci贸n.
Dados los problemas en la medici贸n de la inflaci贸n durante 2007-2015 la base es una construcci贸n propia del CEPXXI
}
