% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexanator.R
\name{indexanator}
\alias{indexanator}
\title{Genera un índice sobre cada variable seleccionada}
\usage{
indexanator(
  data,
  base_indice,
  variables_indexar,
  variables_agrupar,
  pisar_datos = F
)
}
\arguments{
\item{data}{Dataframe que se quiere deflactar/inflar}

\item{base_indice}{Indica sobre que valor se quiere indexar el dato: min, max o fecha YYYY-MM-DD}

\item{variables_indexar}{Vector con los nombres de las variables que se quieren deflactar}

\item{variables_agrupar}{Nombre de la variable que indica el mes en la base a deflactar. Dicha variable debe estar en formato "YYYY-MM-DD"}

\item{pisar_datos}{Valor T/F que permite pisar las variables que se deflactaron}
}
\value{
A matrix of the infile
}
\description{
Sirve para generar un índice en base 100 según el valor del índice seleccionado
La relatividad de los datos puede ser en función del valor máximo, mínimo o de un mes en particular
}
