% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpia_cuits.R
\name{limpia_cuits}
\alias{limpia_cuits}
\title{Eliminar o indicar cuits publicos y no utiles}
\usage{
limpia_cuits(
  data,
  elimina_cajas = T,
  indica_cajas = F,
  elimina_publico = F,
  indica_publico = T,
  elimina_pub_no_prod = F
)
}
\arguments{
\item{data}{Nombre del elemento al que se quiere añadir las etiquetas}

\item{elimina_cajas}{Se sacan las cajas provinciales y otros cuits no utilizados. Default == T}

\item{indica_cajas}{Indica si el CUIT refiere a una caja o no. Default == F, ya que se suelen eliminar}

\item{elimina_publico}{Se sacan los cuits publicos. Default == F}

\item{indica_publico}{Indica si el CUIT es publico o no y el tipo de CUIT público. Inclusion 1 indica que es público productivo, 0 que no lo es y 2 que no es público}

\item{elimina_pub_no_prod}{Se sacan los cuits públicos no productivos (inclusion == 0). Default == F}
}
\value{
A matrix of the infile
}
\description{
Permite identificar rápidamente cuits públicos o de cajas provinciales para indicar cuales son o eliminarlos de la base
}
