% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mectra_demografia.R
\name{mectra_demografia}
\alias{mectra_demografia}
\title{Añade año de nacimiento y sexo al nacer segun CUIL}
\usage{
mectra_demografia(data, sexo = T, nacimiento = T)
}
\arguments{
\item{data}{Nombre del elemento al que se quiere añadir las etiquetas}

\item{sexo}{Calcula el sexo al nacer. Genera la variable sexo_calc. Toma valor 1 si es hombre . Default T}

\item{nacimiento}{Calcula el año de nacimiento. Default T}
}
\value{
A matrix of the infile
}
\description{
Standariza la forma de añadir el año de nacimiento y el sexo al nacer de los que no tenemos en el padron
}
