% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mectra_numerica.R
\name{mectra_numerica}
\alias{mectra_numerica}
\title{Convierte todas las variables numericas de mectra en double}
\usage{
mectra_numerica(data)
}
\arguments{
\item{data}{Nombre del elemento al que se quiere añadir las etiquetas}
}
\value{
A matrix of the infile
}
\description{
Hace más rápido el proceso de pasar los integer y las string que deberian ser double a double
}
